consolidate assign_wcs functions to one module.
